# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateContactRequest:
    properties:
      company_name:
        maxLength: 255
        type: string
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        maxLength: 100
        type: string
      phone_number:
        maxLength: 20
        type: string
      status:
        enum:
        - Active
        - Inactive
        type: string
    required:
    - email
    - first_name
    - status
    type: object
  dto.UpdateContactRequest:
    properties:
      company_name:
        maxLength: 255
        type: string
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        maxLength: 100
        type: string
      phone_number:
        maxLength: 20
        type: string
      status:
        enum:
        - Active
        - Inactive
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing contacts.
  title: Contact Management API
  version: "1.0"
paths:
  /contacts:
    get:
      description: Fetch contacts with pagination, filtering and search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Contact status
        in: query
        name: status
        type: string
      - description: Sort field
        in: query
        name: sort_by
        type: string
      - description: Sort order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: List contacts
      tags:
      - Contacts
    post:
      consumes:
      - application/json
      description: Create a new contact
      parameters:
      - description: Contact payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Create contact
      tags:
      - Contacts
  /contacts/{id}:
    delete:
      description: Soft delete contact
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Delete contact
      tags:
      - Contacts
    get:
      description: Fetch single contact details
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get contact by ID
      tags:
      - Contacts
    put:
      consumes:
      - application/json
      description: Update existing contact
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Update contact
      tags:
      - Contacts
swagger: "2.0"